refactor : style -> styles로 변경
