+++ src/components/Register/TOS/TOS.styles.ts
