+++ src/components/Register/TosItems/TosItem.styles.ts
